tenant/requirements.txt
gunicorn
Flask
Flask-Babel
Flask-SQLAlchemy
Flask-WTF
Flask-Bootstrap
WTForms
docker
celery
python-slugify